Just don't use flycheck in haskell i guess
Whenever there's a problem it uses message to display a checker
warning. Fucking WHY? Who knows but its a multi-line error which was
doing my head in. Choices were to either disable messages entirely or
disable flycheck entirely for haskell. Throwing the baby out with the
bath water a bit but it gave me no choice.