Set shell file name to /bin/zsh
This is a really cheesy fix but my local machine and and the machines I
remote into use this path. It seems like remote commands sometimes use
the local machine's shell path to execute on the remote so this works
around that issue. I can now use dumb-jump on my own machine and on my
remote machines.